requirements.txt
python-benedict[yaml] >= 0.25.0
prompt_toolkit >= 3.0.28